requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
pylint==2.15.3
requests
openai
transformers
flask 
python-dotenv


